+++ include/madara/expression/CompositeGreaterThanEqualNode.cpp
@@ -77,7 +77,7 @@ madara::expression::CompositeGreaterThanEqualNode::prune (bool & can_change)
 
   can_change = left_child_can_change || right_child_can_change;
 
-  return knowledge::KnowledgeRecord::Integer (left_value >= right_value);
+  return knowledge::KnowledgeRecord (left_value >= right_value);
 }
 
 
@@ -87,7 +87,7 @@ madara::knowledge::KnowledgeRecord
 madara::expression::CompositeGreaterThanEqualNode::evaluate (
   const madara::knowledge::KnowledgeUpdateSettings & settings)
 {
-  return knowledge::KnowledgeRecord::Integer (
+  return knowledge::KnowledgeRecord (
     left_->evaluate (settings) >= right_->evaluate (settings));
 }
