+++ include/madara/expression/CompositeLessThanNode.cpp
@@ -74,7 +74,7 @@ madara::expression::CompositeLessThanNode::prune (bool & can_change)
 
   can_change = left_child_can_change || right_child_can_change;
 
-  return knowledge::KnowledgeRecord::Integer (
+  return knowledge::KnowledgeRecord (
     left_value < right_value);
 }
 
@@ -84,7 +84,7 @@ madara::knowledge::KnowledgeRecord
 madara::expression::CompositeLessThanNode::evaluate (
   const madara::knowledge::KnowledgeUpdateSettings & settings)
 {
-  return knowledge::KnowledgeRecord::Integer (
+  return knowledge::KnowledgeRecord (
     left_->evaluate (settings) < right_->evaluate (settings));
 }
