+++ include/madara/expression/SystemCallGetClock.cpp
@@ -24,7 +24,7 @@ madara::expression::SystemCallGetClock::~SystemCallGetClock (void)
 madara::knowledge::KnowledgeRecord
 madara::expression::SystemCallGetClock::item (void) const
 {
-  return madara::knowledge::KnowledgeRecord::Integer (nodes_.size ());
+  return madara::knowledge::KnowledgeRecord (nodes_.size ());
 }
 
 /// Prune the tree of unnecessary nodes. 
@@ -68,7 +68,7 @@ const madara::knowledge::KnowledgeUpdateSettings & settings)
     madara_logger_ptr_log (logger_, logger::LOG_MINOR,
       "System call get_clock is returning the clock of its first argument\n");
 
-    return madara::knowledge::KnowledgeRecord::Integer (
+    return madara::knowledge::KnowledgeRecord (
       nodes_[0]->evaluate (settings).clock);
   }
   else
@@ -76,7 +76,7 @@ const madara::knowledge::KnowledgeUpdateSettings & settings)
     madara_logger_ptr_log (logger_, logger::LOG_MINOR,
       "System call get_clock is returning the system clock\n");
 
-    return madara::knowledge::KnowledgeRecord::Integer (
+    return madara::knowledge::KnowledgeRecord (
       context_.get_clock ());
   }
