+++ include/madara/expression/SystemCallReadFile.cpp
@@ -23,7 +23,7 @@ madara::expression::SystemCallReadFile::~SystemCallReadFile (void)
 madara::knowledge::KnowledgeRecord
 madara::expression::SystemCallReadFile::item (void) const
 {
-  return madara::knowledge::KnowledgeRecord::Integer (nodes_.size ());
+  return madara::knowledge::KnowledgeRecord (nodes_.size ());
 }
 
 /// Prune the tree of unnecessary nodes.
